Revert "up"
This reverts commit 55f863ee21d81909f8a0faecc1bbe094dcfcc155.

+++ plugin/src/main/groovy/com/gmail/arthurstrokov/plugin/tasks/MinorReleaseTask.groovy
@@ -11,7 +11,7 @@ class MinorReleaseTask extends DefaultTask {
     def minorRelease() {
         def currentTagVersion = GitTagVersionService.gitTagVersion()
         def tagVersionDetail = currentTagVersion.split('\\.')
-        def updateMinorTagVersion = Float.parseFloat(tagVersionDetail[1]) + 1
+        def updateMinorTagVersion = Integer.parseInt(tagVersionDetail[1]) + 1
         def newTagVersion = String.join(".", tagVersionDetail[0], updateMinorTagVersion as String)
         GitCommandService.addNewTag(newTagVersion)
         ("git push origin $newTagVersion").execute()
